added script to generate config and pre-push hook for git
