+++ conf_templates/webserv.conf
@@ -1,0 +1,56 @@
+#default conf
+
+server {
+    listen       127.0.0.1:8080;
+
+    #comment;
+	#
+    client_max_body_size    8192;
+
+    location / {
+        alias 					$WWW/html/;
+        index                   index.html index.htm;
+        allow_method            GET POST PUT;
+        auth_basic              "Access to the Webserver";
+        auth_basic_user_file    $PWD/.htpasswd;
+		cgi	.php;
+		cgi_path  /usr/bin/php-cgi;
+    }
+
+    location /webserv {
+        alias                   $WWW;
+        auth_basic              "Access to the Webserver";
+        upload_store            $WWW/testing_put/;
+        auth_basic_user_file    $PWD/conf/.htpasswd;
+        location_max_body_size  4096;
+    }
+
+    location /ok {
+        alias   $WWW/yes/html;
+        index  index.html index.htm;
+		cgi	.php;
+		cgi_path  /usr/bin/php-cgi;
+    }
+
+
+    location /images {
+        alias $WWW/yes/html;
+        index  index.html index.htm;
+		#cgi	.php;
+		#cgi_path  /path/to/exec;
+    }
+
+    location /images/ok/non {
+        alias $WWW/yes/html;
+        index  index.html index.htm;
+		#cgi	.php;
+		#cgi_path  /path/to/exec;
+    }
+
+    location /path/to {
+        alias $WWW/other/html;
+        index  index.html index.htm;
+		cgi	.php;
+		cgi_path  /usr/bin/php-cgi;
+    }
+}
