--- test_conf/conf4_error.conf
@@ -1,40 +1,0 @@
-#default conf
-
-#comment
-
-server {
-    listen       5000;
-    server_name  localhost
-
-    #comment;
-
-
-    location / {
-        root   /usr/share/nginx/html;
-        index  index.html index.htm;
-    }
-
-    # redirect server error pages to the static page /50x.html
-    #
-    error_page   500 502 503 504  /50x.html;
-    }
-
-#another server
-
-server {
-	client_max_body_size 69;
-	listen 5000;
-	server_name hello;
-	error_pages	418 /blabla;
-
-	location /{
-		index index.html;
-		limit_except GET POST HEAD;
-		root	/hello;
-		autoindex off;
-		upload_store    /path/to/upload;
-	}
-
-
-}
-
