--- webserv.conf
@@ -1,210 +1,0 @@
-#default conf
-
-server {
-    listen       127.0.0.1:8080;
-
-    #comment;
-	#
-    client_max_body_size    8192;
-
-    location / {
-        alias 					/usr/share/nginx/yes/html/;
-        index                   index.html index.htm;
-        allow_method            GET POST PUT;
-        auth_basic              "Access to the Webserver";
-        auth_basic_user_file    test_conf/.htpasswd;
-		cgi	.php;
-		cgi_path  /usr/bin/php-cgi;
-    }
-
-    location /webserv {
-        alias                   /Users/wbarendr/Documents/webserv;
-        auth_basic              "Access to the Webserver";
-        upload_store            /Users/wbarendr/Documents/webserv/testing_put/;
-        auth_basic_user_file    test_conf/.htpasswd;
-        location_max_body_size  4096;
-    }
-
-    location /ok {
-        alias   /usr/share/nginx/yes/html;
-        index  index.html index.htm;
-		cgi	.php;
-		cgi_path  /usr/bin/php-cgi;
-    }
-
-
-    location /images {
-        alias /usr/share/nginx/yes/html;
-        index  index.html index.htm;
-		#cgi	.php;
-		#cgi_path  /path/to/exec;
-    }
-
-    location /images/ok/non {
-        alias /usr/share/nginx/yes/html;
-        index  index.html index.htm;
-		#cgi	.php;
-		#cgi_path  /path/to/exec;
-    }
-
-    location /path/to {
-        alias /usr/share/nginx/other/html;
-        index  index.html index.htm;
-		cgi	.php;
-		cgi_path  /usr/bin/php-cgi;
-    }
-
-#     }
-
-#server {
-#    listen       127.0.0.2:8080;
-#
-#    #comment;
-#
-#
-#    location /images {
-#        alias   /usr/share/nginx/other/html;
-#        index  index.html index.htm;
-#		#cgi	.php;
-#		#cgi_path  /path/to/exec;
-#    }
-#}
-#
-#server {
-#    listen       127.0.0.3:5000;
-#	server_name	cheese;
-#
-#    #comment;
-#
-#
-#    location /images {
-#        alias   /usr/share/nginx/cheese/html;
-#        index  index.html index.htm;
-#		#cgi	.php;
-#		#cgi_path  /path/to/exec;
-#    }
-#
-#    }
-#
-#server {
-#    listen       127.0.0.3:5000;
-#	server_name	yes;
-#
-#    #comment;
-#
-#
-#    location /images {
-#        alias   /usr/share/nginx/cheese/html;
-#        index  index.html index.htm;
-#		#cgi	.php;
-#		#cgi_path  /path/to/exec;
-#    }
-#
-#    }
-#
-#server {
-#    listen       127.0.0.3:5000; 
-#	server_name	no;
-#
-#    #comment;
-#
-#
-#    location /images {
-#        alias   /usr/share/nginx/cheese/html;
-#        index  index.html index.htm;
-#		#cgi	.php;
-#		#cgi_path  /path/to/exec;
-#    }
-#
-#    }
-#
-#
-#
-#server {
-#    listen       5000;
-#
-#    #comment;
-#
-#
-#    location / {
-#        alias   /usr/share/nginx/html;
-#        index  index.html index.htm;
-#    }
-#
-#    }
-#
-#server {
-#    listen       5000;
-#	server_name	one;
-#
-#    #comment;
-#
-#
-#    location / {
-#        alias   /usr/share/nginx/html; 
-#        index  index.html index.htm;
-#    }
-#
-#    }
-#
-#server {
-#    listen       5000;
-#	server_name	two;
-#
-#    #comment;
-#
-#
-#    location / {
-#        alias   /usr/share/nginx/html; 
-#        index  index.html index.htm;
-#    }
-#
-#    }
-#
-#server {
-#    listen       8080;
-#
-#    #comment;
-#
-#
-#    location / {
-#        alias   /usr/share/nginx/html; 
-#        index  index.html index.htm;
-#    }
-#
-#    }
-
-
-#server {
-#	listen 1234;
-#	location /oi {
-#		alias /path;
-#	}
-#}
-#
-#server {
-#	listen 8080;
-#	location /oi {
-#		alias /path;
-#	}
-#}
-#
-#another server
-
-#server {
-#	client_max_body_size 69;
-#	listen 5000;
-#	server_name hello;
-#	error_pages	418 /blabla;
-#
-#	location /{
-#		index index.html;
-#		limit_except GET POST HEAD;
-#		alias	/hello;
-#		autoindex off;
-#		upload_store    /path/to/upload;
-#	}
-#
-#
-#}
-#
